requirement.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-environ==0.12.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
